+++ orcamentos-sinapi-application/src/main/java/com/branches/cpu/model/ItemOrcamento.java
@@ -13,7 +13,7 @@ public class ItemOrcamento {
     @EqualsAndHashCode.Include
     private Long id;
     private Insumo insumo;
-    private Double quantidade;
+    private Integer quantidade;
     private Orcamento orcamento;
     private Double valorTotal;
